+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Encrypted data file with passphrase prompts on each run.
+- `version` command to print the CLI version.
 
 ## [v0.4.0] - 2026-01-04